requirements.txt
Flask
openai
PyMuPDF
python-dotenv
python-docx
tiktoken